Fix check for edit values in range
Since the start_time virtual field requires there to be a value part of
start/end time, it throws an exception when no value is there for the
calculation. We need to check that the parts exists themselves. Since
seconds is the value that will always be present we test on that. This
is temporary however and the different parts should be changed to a
singular float second value just like the YouTube API itself uses.