Add Zoom JS class
Its main purpose is to encapsulate zoom behaviour in the client-side:
converting and reverting points from/to a zoom level so that the
start/end values from the range inputs are proportional to the actual
duration of the video. We also need to revert this when loading a
section for edition that has zooms, and therefore has an active zoom
level. Since we persist values that are real (in the sense that are
actual points in time related to the actual duration in the video) in
order to loop the real values we must restore the real points to their
equivalent at the active zoom level.