Add Timeline class
Represents the whole timeline for the video. It depends on initial and
ending values that represent the time spanning in the video. It
generates a timeline following the following rules:

- Minimum 10 labeled ticks
- 4 unlabeled ticks between labeled ticks
- Labels must correspond to half or full-minute values only

Except for values that are too short to hold the rules, in which case we
just use a constant period.