Add timeline to Section
We lazily create and memoize a timeline for the section based on the
operating range of the section. The operating range is the section of
the whole duration that actions can be taken in for this section. In the
case of a zoomed-in section it will be the active zoom range or the
default duration if not.