Update timeline around zoom actions
This is a massive refactor of zoom functionality to move as much of the
logic to the backend via turbo frames and streams. Since we need to
replace the timeline after each zoom action as a means to clearly
communicate to the user that the span of time we're operating on for
picking a point has changed, we take this chance to move actions related
to turbo stream responses.

The main logic of this refactor is to replace JS for the Zoom Out
action with turbo streams via a form in the same vein as the one for the
Zoom In action. We also extract controllers for both action's forms so
that their responsibilities are to keep them up to date with the values
they require for submission based on range and zoom changes. We use the
Zoom Manager-provided state to drive the decisions in both form
controllers, thus centralizing the state, via an outlet. It should be
fine because we don't mutate the state at any point. We should perhaps
enforce this more strongly later on.

The e2e specs remained mostly unchanged except for the obvious need to
change the way we detect the zoom-out element. This should probably be
improved to not have to depend on classes, but the whole test suit
suffers from that so we leave it for when the time comes to make the
whole test suite reusable.