Support Section form submission failure
To submit the Section form we ask the Range stimulus controller to set
the range input's values to full-duration values e.g. from zoomed-in
values. If we react to submission error (like validation errors) we need
to restore the inputs to their previous operational values to allow the
user to continue to work on the range.

- In the Range controller: Include the values to restore the range input
  in the payload of the submitForm event when zoomed-in from the Range
  controller
- In the Range controller: add a new event that restores the range input
  from the restore payload values
- In the Section Form controller: Add a listener to turbo submission
  event to determine if the submission failed or not by checking the
  events related to the form itself, which is the subject element of the
  controller. On error use the restore payload received from Range
  Controller to dispatch the event for it to perform the restoration